Fix handling of SQL errors during reporting
`psql` only returns non-zero in the case of a script error if
`ON_ERROR_STOP` is set.  We rely on a non-zero return from `psql` to
detect that the report did not run correctly so as to e.g. avoid
sending an email with an empty attachment.

Prior to this commit, this handling worked correctly in the event
`psql` failed due to an internal error or due to a failed connection
to the server, but it did not work if there was an error in the SQL
script itself.

With this commit, `psql` will return non-zero to us on script errors
so we can handle them appropriately.